refactor: :recycle: replace LoginModal with AuthModal and update header component for authentication handling
